+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - mark messages as seen on IMAP in batches #3223
 - remove Received: based draft detection heuristic #3230
 - Use pkgconfig for building Python package #2590
+- Do not assign group IDs to ad-hoc groups #2798
 
 
 ## 1.77.0